*로딩화면 구현* 1. 데이터를 받아오는 동안 로딩 페이지를 보여주도록 구현

+++ lib/screens/weather_screen.dart
@@ -39,8 +39,8 @@ class _WeatherScreenState extends State<WeatherScreen> {
         int condition=weatherData['weather'][0]['id'];
         int index=airData['list'][0]['main']['aqi'];
         des=weatherData['weather'][0]['description'];
-        dust1=airData['list'][0]['components']['pm10'];
-        dust2=airData['list'][0]['components']['pm2_5'];
+        dust1=airData['list'][0]['components']['pm10'].toDouble();
+        dust2=airData['list'][0]['components']['pm2_5'].toDouble();
         temp=temp2.round();
         cityName=weatherData['name'];
         icon=model.getWeatherIcon(condition);
